// src/app/constants/interfaces.ts
export interface CommunityInterface {
    CommunityId: number;
    CommunityName: string;
    CommunityDescription: string;
    DateCreated: string;
    Latitude: string;
    Longitude: string;
    City: string;
    State: string;
    Country: string;
    CommunityCode: string;
    CommunitySize: number;
    IsPublic: boolean;
    IsActive: boolean;
    TimeZoneId: number;
    TimeZoneName: string;
    TimeWithOffset: string;
    TimeZoneAbbr: string;
  }
  export interface NearByMatchpointsInteface {
    AddressLine: string;
    ApprovedByPersonId: number;
    CancelPublish: boolean;
    City: string;
    CountryRegion: string;
    CreatedByPersonId: number;
    DateApproved: string;
    DateCreated: string;
    DateModified: string;
    DateSubmitted: string;
    Deleted: boolean;
    DependentCity: string;
    DistanceMeters: number;
    Easting: number;
    GridCellId: number;
    IsMain: boolean;
    IsOnDemand: boolean;
    Latitude: number;
    LocationDescr: string;
    LocationDescription: string;
    LocationId: number;
    LocationName: string;
    LocationOfficialName: string;
    LocationPicture1Id: number;
    LocationPicture2Id: number;
    LocationTypeId: number;
    Longitude: number;
    LoweredUserName: string;
    MatchPointRating: number;
    ModifiedByPersonId: number;
    NightLighting: boolean;
    Northing: number;
    PaidParkingAllowed: boolean;
    Photo: string;
    PostalCode: string;
    PubLocationId: number;
    PubLocationStatus: number;
    PubLocationStatusReason: string;
    PubTransStopGermanyFilteredId: number;
    Publish: boolean;
    RegStopNo: number;
    ShortTermParkingAllowed: boolean;
    State: string;
    Status: string;
    StoppingAllowed: boolean;
    UtmHemisphere: string;
    UtmZone: number;
    WhereToStand: string;
  }